chore(kern): move tcb into own mod
